Rename classes indicating BFT direction
