+++ radixdlt/src/main/java/com/radixdlt/CerberusModule.java
@@ -22,10 +22,10 @@ import com.google.inject.Provides;
 import com.google.inject.Scopes;
 import com.google.inject.Singleton;
 import com.radixdlt.consensus.Consensus;
-import com.radixdlt.consensus.tempo.Application;
+import com.radixdlt.consensus.tempo.MemPool;
 import com.radixdlt.consensus.tempo.Scheduler;
 import com.radixdlt.consensus.tempo.SingleThreadedScheduler;
-import com.radixdlt.consensus.tempo.Tempo;
+import com.radixdlt.consensus.tempo.ChainedBFT;
 import com.radixdlt.consensus.tempo.WallclockTimeSupplier;
 import com.radixdlt.engine.RadixEngine;
 import com.radixdlt.middleware2.converters.AtomToBinaryConverter;
@@ -36,15 +36,15 @@ import com.radixdlt.store.LedgerEntryStore;
 import org.radix.time.Time;
 
 // FIXME: static dependency on Time
-public class TempoModule extends AbstractModule {
+public class CerberusModule extends AbstractModule {
 	@Override
 	protected void configure() {
 		// dependencies
 		bind(Scheduler.class).toProvider(SingleThreadedScheduler::new);
 		bind(WallclockTimeSupplier.class).toInstance(Time::currentTimestamp);
-		bind(Consensus.class).to(Tempo.class).in(Scopes.SINGLETON);
+		bind(Consensus.class).to(ChainedBFT.class).in(Scopes.SINGLETON);
 
-		bind(Application.class).to(RadixEngineAtomProcessor.class);
+		bind(MemPool.class).to(RadixEngineAtomProcessor.class);
 	}
 
 	// We want to use the same instance for Application and RadixEngineAtomProcessor
