+++ radixdlt/src/main/java/com/radixdlt/consensus/tempo/MemPool.java
@@ -22,6 +22,6 @@ import com.radixdlt.store.LedgerEntry;
 /**
  * This is a temporary, rough interface representing the application-side part of the currently developing app/consensus interface.
  */
-public interface Application {
+public interface MemPool {
 	LedgerEntry takeNextEntry() throws InterruptedException;
 }
